IExpressionAccumulator
