+++ Assets/VRM10/Runtime/Components/ExpressionAccumulator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3960aa2fa84da6a46a559ffbc6286cb0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
